check posts swiper
